* Created branch 'update/opt_neTe.py'
* Updated opt_neTe.py
- Beautified code by creating the Optimizer class.
This simplifies the flow of code significantly.
- Changed multiprocessing to send the ne array to the
processpool instead of sending different Monte Carlo iterations
to the pool.
- Removed constraints from the minimization algorithm.
Counterintuitively, this improved the precision of finding
solutions, while also increasing the speed.
- Removed Jacobian evaluation from the minimization algorithm.
This also turned out to only slow down the code while also
decreasing the precision.
- The minimization is now done without constraints, and
the non-physical solutions are removed after a minimum is found.
- Added the used bias coefficients to the output file.
- The code now generates a plot of the used bias coefficients
during runtime.